fix(bookings): constrain and widen the booking columns
- Email and name: required, bounded, validated
- Id: widened to long
- Migrations: ConstrainBookingUserFields, WidenBookingId

+++ Coworking/src/Coworking.API/Models/Responces/CreateBookingResponse.cs
@@ -3,5 +3,5 @@
 public record CreateBookingResponse
 (
     Guid AccessCode,
-    int BookingId
+    long BookingId
 );
